requirements.txt
streamlit>=1.32.0
sentence-transformers>=2.6.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0
plotly>=5.20.0
pandas>=2.1.0
reportlab>=4.1.0
